backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
google-genai>=0.3.0
deep-translator==1.11.4
geopy==2.4.1
python-dotenv==1.0.1
python-multipart==0.0.9
pandas==2.2.2
Pillow==10.3.0
requests==2.32.3
firebase-admin==6.5.0
firebase-functions==0.2.0
a2wsgi==1.10.4
